modules/mlflow/mlflow-image/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.202
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.3
    # via aws-cdk-lib
aws-cdk-integ-tests-alpha==2.149.0a0
    # via -r requirements.in
aws-cdk-lib==2.149.0
    # via
    #   -r requirements.in
    #   aws-cdk-integ-tests-alpha
    #   cdk-ecr-deployment
    #   cdk-nag
boto3==1.34.35
    # via -r requirements.in
botocore==1.34.40
    # via
    #   boto3
    #   s3transfer
cattrs==23.2.3
    # via jsii
cdk-ecr-deployment==3.0.42
    # via -r requirements.in
cdk-nag==2.28.27
    # via -r requirements.in
constructs==10.3.0
    # via
    #   aws-cdk-integ-tests-alpha
    #   aws-cdk-lib
    #   cdk-ecr-deployment
    #   cdk-nag
importlib-resources==6.1.1
    # via jsii
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.101.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-integ-tests-alpha
    #   aws-cdk-lib
    #   cdk-ecr-deployment
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-integ-tests-alpha
    #   aws-cdk-lib
    #   cdk-ecr-deployment
    #   cdk-nag
    #   constructs
    #   jsii
pydantic==2.7.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
python-dotenv==1.0.1
    # via pydantic-settings
s3transfer==0.10.0
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-integ-tests-alpha
    #   aws-cdk-lib
    #   cdk-ecr-deployment
    #   cdk-nag
    #   constructs
    #   jsii
typing-extensions==4.9.0
    # via
    #   jsii
    #   pydantic
    #   pydantic-core
urllib3==1.26.19
    # via botocore
